requirements.txt
# ── Telegram Bot ─────────────────────────────────────────────────────────────
python-telegram-bot==21.6
python-dotenv==1.0.1

# ── HTTP client (for calling the classifier API) ──────────────────────────────
httpx==0.27.0

# ── No other runtime deps needed — SQLite is stdlib ──────────────────────────
